pyproject.toml
[project]
name = "nonebot-plugin-remind"
version = "0.1.0"
description = "nonebot2插件：提供符合中国宝宝体质的定时提醒功能~"
authors = [
    {name = "Elden",email = "h_elden@qq.com"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "nonebot2 (>=2.3.0,<3.0.0)",
    "nonebot-plugin-localstore (>=0.7.3,<0.8.0)",
    "nonebot-adapter-onebot (>=2.4.6,<3.0.0)",
    "nonebot-plugin-apscheduler (>=0.5.0,<0.6.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
